+++ packages/tools/atlas/src/bin/base/NUTS/13-4_getCountriesTopojson.js
@@ -16,6 +16,12 @@ import {
 	NUTS_DATABASE_DIR_4,
 } from '../../../lib/paths.js';
 
+/*
+- read countries_by_year.yaml
+- read topojson files
+- for all countries, filter topojsons by CNTR_CODE and save them
+*/
+
 /* paths */
 
 const inPaths = {
@@ -43,17 +49,6 @@ const makeTopojsonFilterByCountryId = countryId => transformPaths({
 
 /* run */
 
-/*
-- read countries_by_year.js
-- read topojson files
-- for all countries, filter topojsons by CNTR_CODE and save them
-
-in:
-	- NUTS_DATABASE_DIR_1/countries_by_year.js
-	- NUTS_DATABASE_DIR_1/topojson/*.json
-out:
-	- NUTS_DATABASE_DIR_2/topojson/*_{country}.json
-*/
 const run = async () => {
 	const filtersByYear =
 		await import(inPaths.countriesByYear)
@@ -70,29 +65,28 @@ const run = async () => {
 
 	await Promise.all(
 		_.flatMap(filenames, async filename => {
-			const inPath = path.resolve(inDirs.topojson, filename);
+			const topoJsInPath = path.resolve(inDirs.topojson, filename);
 			const {name} = path.parse(filename);
 			const [ , , , year] = name.split('_');
 
-			const {default: topojson} = await import(inPath);
+			const {default: topojson} = await import(topoJsInPath);
 			const {length} = JSON.stringify(topojson);
 
 			return _.map(filtersByYear[year], ([id, filterByCountryId]) => {
-				const outPath = path.resolve(outDirs.topojson, `${name}_${id}.js`);
+				const topoJsOutPath = path.resolve(outDirs.topojson, `${name}_${id}.js`);
 
 				const filteredObjectsTopo = filterByCountryId(topojson);
 				const prunedTopo = pruneTopology(filteredObjectsTopo);
 
-				return saveExportedObj(outPath)(prunedTopo)
-				.then(tapMessage(`${year} ${id}: ${length} -> ${JSON.stringify(prunedTopo).length} Saved in ${outPath}`))
-				.catch(err => console.error(outPath, err));
+				return saveExportedObj(topoJsOutPath)(prunedTopo)
+				.then(tapMessage(`${year} ${id}: ${length} -> ${JSON.stringify(prunedTopo).length} Saved in ${topoJsOutPath}`))
+				.catch(err => console.error(topoJsOutPath, err));
 			});
 		})
 	)
 }
 
 console.log(`\nrun: ${path.basename(import.meta.url)}\n`);
-console.log('Fetching, please wait...');
 
 run()
 .then(tapMessage('Done'))