+++ packages/tools/atlas/src/lib/nutsUtils.js
@@ -1,3 +1,4 @@
+import {makeFilterTopoBy} from '@svizzle/geo';
 import * as _ from 'lamb';
 
 export const getNutsName = _.adapter([
@@ -5,3 +6,8 @@ export const getNutsName = _.adapter([
 	_.casus(_.hasKey('NUTS_NAME'), _.getKey('NUTS_NAME')),
 	_.always('Unknown'),
 ]);
+
+export const makeFilterTopojsonByNutsId = makeFilterTopoBy({
+	objKey: 'NUTS',
+	propKey: 'NUTS_ID',
+});